minor update zemu tests
